fix(core) ensure absolute path is used for event sock
This fixes an issue with how Kong initializes resty.events. The code was
previously using `ngx.config.prefix()` to determine the listening socket
path to provide to the resty.events module. This causes breakage when
NGINX is started with a relative path prefix:

> failed to instantiate 'kong.worker_events' module: failed to disable listening: unix:my-prefix/worker_events.sock

Instead of using `ngx.config.prefix()`, Kong will now prefer
`kong.configuration.prefix` when available, as it is already normalized
to an absolute path. If `kong.configuration.prefix` is not defined, the
result of `ngx.config.prefix()` will be used after resolving it to an
absolute path.